Move Settings utils to printrun/settings.py
